+++ repertoire.server/test/integration/song/section/create_song_section_test.go
@@ -47,7 +47,7 @@ func assertCreatedSongSection(
 	assert.Equal(t, request.SongID, songSection.SongID)
 	assert.Equal(t, request.Name, songSection.Name)
 	assert.Equal(t, request.TypeID, songSection.SongSectionTypeID)
-	assert.Equal(t, uint(model.DefaultSongSectionConfidence), songSection.Confidence)
+	assert.Equal(t, model.DefaultSongSectionConfidence, songSection.Confidence)
 	assert.Zero(t, songSection.Rehearsals)
 	assert.Equal(t, uint(order), songSection.Order)
 }
