+++ repertoire.server/test/unit/domain/usecase/song/create_song_test.go
@@ -270,7 +270,7 @@ func assertCreatedSong(
 	for i, section := range request.Sections {
 		assert.NotEmpty(t, song.Sections[i].ID)
 		assert.Equal(t, section.Name, song.Sections[i].Name)
-		assert.Equal(t, uint(model.DefaultSongSectionConfidence), song.Sections[i].Confidence)
+		assert.Equal(t, model.DefaultSongSectionConfidence, song.Sections[i].Confidence)
 		assert.Zero(t, song.Sections[i].Rehearsals)
 		assert.Equal(t, uint(i), song.Sections[i].Order)
 		assert.Equal(t, section.TypeID, song.Sections[i].SongSectionTypeID)
